--- a/ecc/bw6-761/bw6-761.go
+++ b/ecc/bw6-761/bw6-761.go
@@ -41,6 +41,9 @@ var bCurveCoeff fp.Element
 // bTwistCurveCoeff b coeff of the twist (defined over Fp) curve
 var bTwistCurveCoeff fp.Element
 
+// twoInv 1/2 mod p (needed for DoubleStep in Miller loop)
+var twoInv fp.Element
+
 // generators of the r-torsion group, resp. in ker(pi-id), ker(Tr)
 var g1Gen G1Jac
 var g2Gen G2Jac
@@ -56,7 +59,7 @@ var g2Infinity G2Jac
 // Miller loop 1: f(P), div(f) = (x+1)(Q)-([x+1]Q)-x(O)
 // Miller loop 2: f(P), div(f) = (x**3-x**2-x)(Q) -([x**3-x**2-x]Q)-(x**3-x**2-x-1)(O)
 var loopCounter1 [64]int8
-var loopCounter2 [127]int8
+var loopCounter2 [190]int8
 
 // Parameters useful for the GLV scalar multiplication. The third roots define the
 //  endomorphisms phi1 and phi2 for <G1Affine> and <G2Affine>. lambda is such that <r, phi-lambda> lies above
@@ -79,6 +82,8 @@ func init() {
 	bCurveCoeff.SetOne().Neg(&bCurveCoeff)
 	bTwistCurveCoeff.SetUint64(4)
 
+	twoInv.SetOne().Double(&twoInv).Inverse(&twoInv)
+
 	g1Gen.X.SetString("6238772257594679368032145693622812838779005809760824733138787810501188623461307351759238099287535516224314149266511977132140828635950940021790489507611754366317801811090811367945064510304504157188661901055903167026722666149426237")
 	g1Gen.Y.SetString("2101735126520897423911504562215834951148127555913367997162789335052900271653517958562461315794228241561913734371411178226936527683203879553093934185950470971848972085321797958124416462268292467002957525517188485984766314758624099")
 	g1Gen.Z.SetString("1")
@@ -90,10 +95,12 @@ func init() {
 	g1GenAff.FromJacobian(&g1Gen)
 	g2GenAff.FromJacobian(&g2Gen)
 
-	//binary decomposition of 9586122913090633729, little endian
-	loopCounter1 = [64]int8{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1}
+	//binary decomposition of 9586122913090633730, little endian
+	// xGen+1
+	loopCounter1 = [64]int8{0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1}
 
-	T, _ := new(big.Int).SetString("91893752504881257691937156713741811711", 10)
+	// xGen^3-xGen^2-xGen
+	T, _ := new(big.Int).SetString("880904806456922042166256752416502360955572640081583800319", 10)
 	ecc.NafDecomposition(T, loopCounter2[:])
 
 	g1Infinity.X.SetOne()