--- a/ecc/bw6-761/utils.go
+++ b/ecc/bw6-761/utils.go
@@ -15,7 +15,10 @@ func GenFp() gopter.Gen {
 	return func(genParams *gopter.GenParameters) *gopter.GenResult {
 		var elmt fp.Element
 		var b [fp.Bytes]byte
-		rand.Read(b[:])
+		_, err := rand.Read(b[:])
+		if err != nil {
+			panic(err)
+		}
 		elmt.SetBytes(b[:])
 
 		genResult := gopter.NewGenResult(elmt, gopter.NoShrinker)
@@ -67,9 +70,11 @@ func GenFr() gopter.Gen {
 	return func(genParams *gopter.GenParameters) *gopter.GenResult {
 		var elmt fr.Element
 		var b [fr.Bytes]byte
-		rand.Read(b[:])
+		_, err := rand.Read(b[:])
+		if err != nil {
+			panic(err)
+		}
 		elmt.SetBytes(b[:])
-
 		genResult := gopter.NewGenResult(elmt, gopter.NoShrinker)
 		return genResult
 	}