… ecc/bw6-761/internal/fptower/e2.go
… ecc/bw6-761/internal/fptower/e3_test.go
… ecc/bw6-761/internal/fptower/e6_test.go
… ecc/bw6-761/internal/fptower/frobenius.go
… ecc/bw6-761/pairing.go