- Miller loop according to [ABLR13](https://eprint.iacr.org/2013/722.pdf) (same as the other curves).
- `Fp6` as a quadratic extension over `Fp3` (instead of a cubic extension over `Fp2`).

### benchcmp vs develop
- pairing
```
BenchmarkPairing-2          3080057       2788557       -9.46%
BenchmarkMultiPairing-2     5761097       4970379       -13.73%
```
- Miller loop
```
benchmark                 old ns/op     new ns/op     delta
BenchmarkMillerLoop-2     1978253       1634283       -17.39%
```
- final exp
```
benchmark                          old ns/op     new ns/op     delta
BenchmarkFinalExponentiation-2     1107366       1123309       +1.44%
```
- `Fp6`
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkE6Add-2             77.4          75.6          -2.33%
BenchmarkE6Sub-2             59.4          55.2          -7.07%
BenchmarkE6Mul-2             3001          2964          -1.23%
BenchmarkE6Cyclosquare-2     1424          1444          +1.40%
BenchmarkE6Square-2          2120          2123          +0.14%
BenchmarkE6Inverse-2         26141         25171         -3.71%
BenchmarkE6Conjugate-2       54.3          19.7          -63.72%
BenchmarkE6Frobenius-2       485           573           +18.14%
BenchmarkE6Expt-2            107347        106262        -1.01%
```